Interface for ChartSaver and stub for PostgresChartSaver
